Lift data loading out of component, expose via props (#13)
Fixes #11

## What was done?

- "Load document" and all related template loading logic has been lifted
out of the component. It's moved out of `src/lib`, and into the "demo
wrapper" code. That's still within this repo, but not packaged and
published in the npm package.
- To load a GOBL document (typically a JSON string) into the editor, the
user of the lib can now provide a `value` property. - When a value is
provided (e.g. different document/text is loaded), the library will
update its internal stores, so both the Monaco editor value is updated,
and any existing envelope is cleared. It attempts to parse a GOBL
envelope, and if successful, stores the envelope contents as the editor
value and separately stores the envelope. This is similar to how
responses from the GOBL WASM actions are processed.
- To enforce a JSON Schema, a new `jsonSchemaURL` property can be used.
Typically, the user of the lib will provide this with a well-known GOBL
schema URL, either based on the context in the UI or the schema of an
existing GOBL document.
- README was updated with new usage instructions for the above mentioned
props.
- Updated the "demo wrapper" code base, used dark color scheme for top
menu bar.